+++ src/index.ts
@@ -1,11 +1,15 @@
 import React from "react";
 import { LanguageSelector as Component } from "./selector";
 import { setLocaleCookie } from "./utils";
-import type { LanguageSelectorProps, LocaleConfig } from "./types";
+import type {
+  LanguageSelectorProps,
+  LocaleConfig,
+  ReloadStrategy,
+} from "./types";
 
 export const LanguageSelector: (
   props: LanguageSelectorProps,
 ) => React.JSX.Element | null = Component;
 
 export { setLocaleCookie };
-export type { LanguageSelectorProps, LocaleConfig };
+export type { LanguageSelectorProps, LocaleConfig, ReloadStrategy };